==> 5.Convert_regional_whiplash_data_to_decadal_data.py <==
#########################################################################################
## Script to convert the regional whiplash.nc files (if each region was run separately) into 
## decadal files for later analysis. Can be edited to convert a CONUS-wide file into 
## decadal files.
## Bryony Louise Puxley
## Last Edited: Friday, July 25th, 2025
## Input: regional whiplash data
## Output: multiple netCDF files of identified whiplash occurrences at a ~ 6 km grid resolution from
## 1915 to 2020, split into 10-year decade periods. If files were regional, they have been
## combined into CONUS-wide.
#########################################################################################
# Import Required Modules
#########################################################################################
import os
import xarray as xr

#########################################################################################
# Convert regional files into decadal periods 
#########################################################################################
Regions = {"WCN", "WCS", "MWN", "MWC", "MWS", "NGP", "SGP", "NGL", "SGL", "NNE", "SNE", "ESE", "WSE"}
time_periods = {'1915_1924','1925_1934','1935_1944','1945_1954', '1955_1964','1965_1974','1975_1984','1985_1994','1995_2004','2005_2014','2015_2020'}

dirname= '/data2/bpuxley/Whiplash/Regional_data'

pathfiles = []
for i in Regions:
        filename = 'whiplashes_%s.nc'%(i)
        pathfile = os.path.join(dirname, filename)
        pathfiles.append(pathfile)

datasets = [xr.open_dataset(f) for f in pathfiles]
print('Read in Data')

#Split whiplash data into 10-year periods
for period in sorted(time_periods):
	print(period)
	data = [ds.sel(time=slice(period[:4]+"-01-01",period[5:]+"-12-31")) for ds in datasets]
	
	print(period+': Started')
	combined = xr.combine_by_coords(data)
	print(period+': Ended')
	
	print('Saving '+period+': Started')
	combined.to_netcdf('/data2/bpuxley/droughts_and_pluvials/droughts_and_pluvials_%s.nc'%(period))
	print('Saving '+period+': Ended')

	del(combined)
