==> 7.Density_Calculation.py <==
#########################################################################################
## Script to combine the multiple density.nc files into one file and calculate the 99th 
## percentile for polygons.
## Bryony Louise Puxley
## Last Edited: Monday, July 28th, 2025 
## Input: decadal density files.
## Output: The 99th percentile of densities for drought-to-pluvial and pluvial-to-drought 
## events to be used to draw the single contour for the polygon outlining the area defined
## as having a whiplash event.
#########################################################################################
# Import Required Modules
#########################################################################################
import numpy as np
import xarray as xr
from datetime import datetime

#########################################################################################
# Periods
#########################################################################################
Years = {"1915_1924", "1925_1934", "1935_1944", "1945_1954", "1955_1964", "1965_1974",
                        "1975_1984", "1985_1994", "1995_2004", "2005_2014", "2015_2020"}

#########################################################################################
# Import Data - load in all files
#########################################################################################
dirname= '/data2/bpuxley/Density/*.nc'

all_densities = xr.open_mfdataset(dirname, combine='by_coords')

#Save file out
#print('Saving all densities file')
#all_densities.to_netcdf('/scratch/bpuxley/Density/density_1915_2020.nc')
#print('File Saved')

#Split into DP and PD
DP_densities = all_densities.DP_density
PD_densities = all_densities.PD_density

#Calculate the 99th percentile of density files to draw the polygon
print('calculating percentile')
DP_perc = np.percentile(DP_densities, q=99)
print('Drought-to-Pluvial calculated')
PD_perc = np.percentile(PD_densities, q=99)
print('Pluvial-to-Drought calculated')

print(f'the 99th percentile for Drought-to-Pluvial Events is {DP_perc}, for Pluvial-to-Drought Events is {PD_perc}, and they were calculated at {datetime.now()}')
